ref: Add sentry sink
Add a few new sinks to allow dogfooding of DDM from within our metrics
client:

1. A middleware to sample metrics
2. A middleware to mirror metrics
3. A middleware to send metrics to sentry
