+++ Cargo.toml
@@ -20,6 +20,8 @@ cadence = { version = "0.29.0", optional = true }
 log = "0.4"
 signal-hook = { version = "0.3.17", optional = true }
 thread_local = { version = "1.1.7", optional = true }
+sentry = { version = "0.32.1", optional = true, default-features = false, features = ["UNSTABLE_metrics"] }
+rand = { version = "0.8.5", features = ["small_rng"] }
 
 [features]
 default = ["cli"]
